requirements.txt
pandas==1.4.4
featurewiz==0.1.87
logzero==1.7.0
matplotlib==3.5.3
mlxtend==0.19.0
numpy==1.19.5
pandas==1.4.4
pycaret== 2.2.2
PyYAML==6.0
scikit_learn==1.1.2
scipy==1.5.4
seaborn==0.11.2
xgboost==1.6.2
streamlit==1.9.0
protobuf~=3.19.0
shap==0.41.0
joblib==1.1.0